Created utils folder
