+++ src/frontend/src/pages/PIAForm/utils/validationUI.ts
@@ -1,0 +1,42 @@
+import { SetStateAction } from 'react';
+import { PiaValidationMessage } from '..';
+
+type ResetUIProps = {
+  setValidationMessages: (value: SetStateAction<PiaValidationMessage>) => void;
+  setIsValidationFailed: (value: SetStateAction<boolean>) => void;
+  setValidationFailedMessage: (value: SetStateAction<string>) => void;
+  doc: Document;
+};
+
+export const resetUI = (props: ResetUIProps) => {
+  const {
+    setValidationMessages,
+    setIsValidationFailed,
+    setValidationFailedMessage,
+    doc,
+  } = props;
+
+  const reset = doc.getElementsByClassName('is-invalid');
+  if (reset) {
+    [...reset].forEach((el) => {
+      el.classList.remove('is-invalid');
+    });
+    setValidationMessages({});
+    setIsValidationFailed(false);
+    setValidationFailedMessage('');
+  }
+  const richText = doc.getElementsByClassName('richText');
+  if (richText) {
+    [...richText].forEach((el) => {
+      el.classList.remove('form-control');
+    });
+  }
+};
+
+export const highlightInvalidField = (formId: string, doc: Document) => {
+  const element = doc.getElementById(formId);
+  if (element) {
+    element.scrollIntoView({ behavior: 'smooth', block: 'center' });
+    element.className += ' is-invalid form-control';
+  }
+};
